split detectRootFiles tests
